Rotate cutout to ROTATION_0 when getting metrics for natural rotation
The method getPossibleMaximumWindowMetrics() is to let the launcher get
possible display info for different device states. It is expected to
have everything in natural rotation for each device state. All other
fields such as bounds, are rotated.

The DisplayCutout was not rotated as it was always the same across all
rotations. With movable cutout effort, extra rotation is needed to make
sure it always get the natural rotation result.

Test: See verify steps in b/321914205
Bug: 321914205
Bug: 302387383
Change-Id: Ibd7d300b3138714b6d60b6a34a16a9020962ae68