Move SettingsObserver#mLastEnabled to UserData
In order to make InputMethodManagerService.SettingsObserver multi-user
aware, one challenge is

  SettingsObserver#mLastEnabled,

which is tightly coupled with "single-user" concept. So this CL moves
it to UserData object in a mechanical manner.

There must be no behavior change in the single user mode anyway, and
for concurrent multi-user mode SettingsObserver is not yet fully
multi-user aware so this change is not yet used.

Bug: 342027196
Test: presubmit
Flag: android.view.inputmethod.concurrent_input_methods
Change-Id: Ie37a439fcd4155272efa9620ca20520ba4c58400